zzogl: fix wrong type access cmake: don't strip by default the binary in release mode gsdx: disable the regeneration of ogl shader
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@5806 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ plugins/zzogl-pg-cg/opengl/GifTransfer.cpp
@@ -97,7 +97,7 @@ template<int index> void _GSgifTransfer(const u32 *pMem, u32 size)
 
 				if (path->tag.PRE && (path->tag.FLG == GIF_FLG_PACKED))
 				{
-					u32 tagprim = path->tag.PRIM;
+					u64 tagprim = path->tag.PRIM;
 					GIFRegHandlerPRIM((u32*)&tagprim);
 				}
 			}
